refactor(cli): keep attribution lookups in memory
